requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.2.0
pytest-django==4.8.0
sqlparse==0.5.1
